Revert "Clean up global_settings import"
This reverts commit 8e9abb5ea04635f1310678ea14cd4e4d6ad4ad91.
